:truck: Moves Tests to correct files
