+++ size.json
@@ -5,8 +5,8 @@
       "raw": 2161
     },
     "brotli": {
-      "pretty": "904 B",
-      "raw": 904
+      "pretty": "912 B",
+      "raw": 912
     }
   },
   "alpinejs": {
